backend_fastapi/requirements.txt
fastapi==0.115.6
uvicorn==0.30.6
python-dotenv==1.0.1
SQLAlchemy==2.0.36
alembic==1.14.1
psycopg2-binary==2.9.9
pydantic==2.10.3
pydantic-settings==2.6.1
PyJWT==2.10.1
email-validator==2.2.0
